+++ ft_strlcpy.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-int		ft_strlcpy(char *dest, char *str, size_t n)
+size_t	ft_strlcpy(char *dest, char *str, size_t n)
 {
 	int i;
